pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.1",
    "ariadne>=0.26.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.0.0",
    "fastapi>=0.116.1",
    "fastadmin>=0.1.21",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.8.0",
    "redis>=5.0.0",
    "setuptools>=68.0.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
